requirements.txt
absl-py==0.12.0
aniso8601==9.0.1
astunparse==1.6.3
attrs==21.2.0
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
dill==0.3.3
dm-tree==0.1.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
Flask==2.0.1
Flask-RESTful==0.3.9
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.4.0
google-api-core==1.28.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.17.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.19.1
idna==2.10
importlib-resources==5.1.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
murmurhash==1.0.5
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python-headless==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pathy==0.5.2
Pillow==8.2.0
portalocker==2.0.0
preshed==3.0.5
promise==2.3
proto-plus==1.18.1
protobuf==3.17.1
psutil==5.8.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydantic==1.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
pywin32==301
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==1.5.1
scikit-learn==0.24.2
scipy==1.6.3
sentencepiece==0.1.95
seqeval==1.2.2
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.0.0
tensorflow-model-optimization==0.5.0
tensorflow-text==2.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.5.0
tf-slim==1.1.0
thinc==8.0.3
threadpoolctl==2.1.0
tqdm==4.61.0
typeguard==2.12.0
typer==0.3.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
wasabi==0.8.2
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
